exemplo/requirements.txt
Django==3.1.2
django-silk==4.1.0
flake8==3.8.4
graphene-django==2.13.0
graphene==2.1.8